Allowing multiple securityKeys per user
This means that during sign-in, we now send a list of SignRequest
instead of a single SignRequest to the browser. This in turns means
that in the session data, we need to keep track of which public key we
expect the challenge to be signed with. Which necessitates the
splitting up of SessionData into EnrollSessionData (which has no public
key) and SignSessionData (which does).